feat(lungo-fe): render transport from topology message_transport; drop transport config fetch
Render transport node labels and GitHub links from backend message_transport
wire fields and remove the dynamic /transport/config polling path.
