+++ coffeeAGNTCY/coffee_agents/lungo/frontend/src/hooks/useWorkflowGraphTopologySync.ts
@@ -11,16 +11,14 @@ import {
 } from "@/api/agenticWorkflowsClient"
 import type { TopologyWire } from "@/api/agenticWorkflowsTypes"
 import { logger } from "@/utils/logger"
-import { identityUiVariantForPattern } from "@/utils/agenticTopologyIdentityUiMap"
 import { topologyWireToReactFlow } from "@/utils/topologyToReactFlow"
-import type { PatternType } from "@/utils/patternUtils"
 import {
   REFETCH_DEBOUNCE_MS,
   type WorkflowGraphAgenticSession,
 } from "./useWorkflowGraphFromAgenticApi.types"
 
 interface UseWorkflowGraphTopologySyncParams {
-  patternRef: React.RefObject<PatternType>
+  isStreamingRef: React.RefObject<boolean>
   sessionRef: React.RefObject<WorkflowGraphAgenticSession | null>
   onAppliedRef: React.RefObject<(() => void) | undefined>
   attachHandlers: (node: Node) => Node
@@ -30,7 +28,7 @@ interface UseWorkflowGraphTopologySyncParams {
 }
 
 export function useWorkflowGraphTopologySync({
-  patternRef,
+  isStreamingRef,
   sessionRef,
   onAppliedRef,
   attachHandlers,
@@ -44,7 +42,7 @@ export function useWorkflowGraphTopologySync({
     (topology: TopologyWire | undefined) => {
       const { nodes: mappedNodes, edges: mappedEdges } =
         topologyWireToReactFlow(topology, {
-          identityUiVariant: identityUiVariantForPattern(patternRef.current),
+          isStreaming: isStreamingRef.current,
         })
       const withHandlers = mappedNodes.map(attachHandlers)
       lastAppliedGraphNodeIdsRef.current = new Set(
@@ -59,8 +57,8 @@ export function useWorkflowGraphTopologySync({
     },
     [
       attachHandlers,
+      isStreamingRef,
       onAppliedRef,
-      patternRef,
       restoreEdgeAnimation,
       setEdges,
       setNodes,
